move navigation to a separate file
